Move settings_view.py to settings package
Move src/amplifyp/gui/views/settings_view.py to the settings package folder and update imports and tests accordingly.
