+++ src/amplifyp/gui/views/__init__.py
@@ -22,7 +22,7 @@ from .about import AboutView
 from .dimer import DimerView
 from .input import InputView
 from .pcr import PCRView
-from .settings_view import SettingsView
+from .settings import SettingsView
 
 __all__ = [
     "AboutView",
